Added the /calendar and completed visuals
Currently created the calendar page. This is not the completed calendar function. I still need to account for switching months and adding events. However this is a good starting point. I also added a debugging line at the end of our __init__.py file. I thought it would be useful so I left it but we can delete if we want to.