Completed check_and_send_notifications()
Completed and tested the functionaility of the check_and_send_notifications() function. System checks an hour window based on the users preferred notification day window and successfully sends them an email about the event. Verifies if an email has been sent about the event in the past 24 hours and does not send another one until the next day.